start of work to refactor the docs to use readthedocs
this is the first pass to format the previous latex into rst. I have done the conversion and
only started to go through ensuring that content is preserved (meaning I did not miss anything)
and all the code blocks are formatted (after the automated conversion they indeed are not!) I
will want a few more hours to finish this up, and then we need to discuss deployment. E.g., I
recommend readthedocs so you can automate deployment and keep versioned docs. We can also deploy
to github pages (I can make a workflow) but I do not have a good suggestion for versioning things
that way. I also have not added in an ability to still render the pdfs if that is desired, which
I think should be possibly. Finally, it would be nice if some of these docs could render from
docstrings - I know how to do this for Python so I wonder if Cpp is that much different. It is
probably terrible because it is cpp, but what can you do?

Signed-off-by: vsoch <vsoch@users.noreply.github.com>